FormsPreprocessors Reference

FormsPreprocessors.direct_product — Method

direct_product(df, col1, col2, newcol)

Concatenates values of col1 and col2, which is stored in newcol.

Example

julia> df = DataFrame(q1 = ["young", "old", "young", "young"],
    q2 = ["no", "no", "no", "yes"])
julia> direct_product(df, :q1, :q2, :q1xq2)

Row  │ q1      q2      q1xq2   
     │ String  String  String  
─────┼─────────────────────────
1    │ young     no      young_no
2    │ old      no      old_no
3    │ young     no      young_no
4    │ young     yes     young_yes

If either value is missing, the other value itself is stored. If either value is "", the other value with delim is stored.

Example

julia> df = DataFrame(q1 = ["young", missing, "young", ""],
    q2 = ["no", "no", missing, "yes"])
julia> direct_product(df, :q1, :q2, :q1xq2)
4x3 DataFrame
Row  │ q1       q2       q1xq2  
     │ String?  String?  String 
─────┼──────────────────────────
1    │ young      no       young_no
2    │ missing  no       no
3    │ young      missing  young
4    │          yes      _yes

FormsPreprocessors.discretize — Function

discretize(df, key, thresholds, newcol="class_$(String(key))"; newcodes=[])

Classify numerical answers to classes defined by thresholds. The number of classes is length(thresholds) + 1, because the ranges to be (-Inf, thresholds[1]), [thresholds[1], thresholds[2]), ..., [thresholds[end], Inf]. Therefore, the length of newcodes must be length(thresholds) + 1.

Example

julia> df = DataFrame(expense=[100, 250, 300, 1000, 150])
julia> discretize(df, :expense, [100, 300, 500]; newcodes=["No", "Low", "Middle", "High"])
5x2 DataFrame
Row  │ expense  class_expense 
     │ Int64    String        
─────┼────────────────────────
1    │     100  Low
2    │     250  Low
3    │     300  Middle
4    │    1000  High
5    │     150  Low

FormsPreprocessors.onehot — Method

onehot(df, key; ordered_answers=[])

Performs one-hot encoding on key column. The column is expected to be of vectors, as split_mc_col! generates. If you want to sort columns generated, specify ordered_answers. If not ordered_answers specified, columns are ordered by value appearance.

Example

julia> df = DataFrame(item=["Apple;Orange", "Tomato", "Tomato;Pepper"])
julia> onehot(df, :item; ordered_answers=["Tomato", "Pepper", "Apple", "Orange"])
3x5 DataFrame
Row  │ item                  item_Tomato  item_Pepper  item_Apple  item_Orange 
     │ Array…                Any          Any          Any         Any         
─────┼─────────────────────────────────────────────────────────────────────────────────────
1    │ ["Apple", "Orange"]   no           no           yes         yes
2    │ ["Tomato"]            yes          no           no          no
3    │ ["Tomato", "Pepper"]  yes          yes          no          no

FormsPreprocessors.recode — Function

recode(df, key, newkey, vec_from, vec_to=[]; other="other")

Recodes values in key column with values in vec_from into corresponding vec_to values, which are stored in newkey column. Items not in vec_from keep original values. If vec_to is shorter than vec_from, the last values are recoded to other. If you want to recode all irregular answers such as 'other:______' to others, use recode_others. If you want to recode columns with common choices such as choices to matrix type question, use recode_matrix.

Example

julia> df = DataFrame(item = ["Apple", "Orange", "Tomato", "Pepper"])
julia> recode(df, :item, :newitem, ["Apple", "Orange", "Pepper"], ["Fruit", "Fruit"])
4x2 DataFrame
Row  │ item    newitem 
     │ String  String  
─────┼─────────────────
1    │ Apple   Fruit
2    │ Orange  Fruit
3    │ Tomato  Tomato
4    │ Pepper  other

You can recode values stored in a column of vectors which can be generated using split_mc_col!.

!!! Note

If you recode a column of vectors, single-value answer must be vectorized, 
as [`split_mc_col!`](@ref) does.

Example

julia> df = DataFrame(item = [["Apple", "Orange"], ["Tomato"], ["Tomato", "Pepper"]])
julia> recode(df, :item, :newitem, ["Tomato", "Pepper"], ["Vegitable", "Spice"])
3x2 DataFrame
Row  │ item                  newitem                
     │ Array…                Array…                 
─────┼──────────────────────────────────────────────
1    │ ["Apple", "Orange"]   ["Apple", "Orange"]
2    │ ["Tomato"]            ["Vegitable"]
3    │ ["Tomato", "Pepper"]  ["Vegitable", "Spice"]

See also recode_others, recode_matrix

FormsPreprocessors.recode_matrix — Function

recode_matrix(df, keys, vec_from, vec_to=[]; other="other", prefix="r")

Recodes values of vec_from in keys columns to vec_to values. New values from column :foo are stored in corresponding column named :prefix_foo.

Example

julia> df = DataFrame(q1=["Strongly agree", "Disagree", "Agree", "Neutral", "Strongly disagree"], 
    q2 = ["Disagree", "Strongly disagree", "Neutral", "Agree", "Agree"])
julia> recode_matrix(df, [:q1, :q2], ["Strongly agree", "Agree", "Disagree", "Strongly disagree"], 
["t2b", "t2b", "b2b", "b2b"])
5x4 DataFrame
Row  │ q1                 q2                 r_q1     r_q2    
     │ String             String             String   String  
─────┼────────────────────────────────────────────────────────
1    │ Strongly agree     Disagree           t2b      b2b
2    │ Disagree           Strongly disagree  b2b      b2b
3    │ Agree              Neutral            t2b      Neutral
4    │ Neutral            Agree              Neutral  t2b
5    │ Strongly disagree  Agree              b2b      t2b

See also recode

FormsPreprocessors.recode_others — Method

recode_others(df, key, newkey, regular_answers; other="other")

Recodes all appeared values in key column but in regular_answers into other, which are stored in newkey column.

Example

julia> df = DataFrame(item = ["Apple", "Orange", "Tomato", "Pepper"])
julia> recode_others(df, :item, :newitem, ["Apple", "Orange"])

Row  │ item    newitem 
     │ String  String  
─────┼─────────────────
1    │ Apple   Apple
2    │ Orange  Orange
3    │ Tomato  other
4    │ Pepper  other

See also recode

FormsPreprocessors.split_mc_col! — Method

split_mc_col!(df, key; delim=";")

Mutates key columns with values of "delim-concatenated" MA answers into column with vectors.

Example

julia> df = DataFrame(item=["Apple;Orange", "Tomato", "Tomato;Pepper"])
julia> split_mc_col!(df, :item)
3x1 DataFrame
Row  │ item                              
     │ Array…                            
─────┼───────────────────────────────────
1    │ ["Apple", "Orange"]
2    │ ["Tomato"]
3    │ ["Tomato", "Pepper"]